+++ src/dash-table/conditional/index.ts
@@ -4,6 +4,11 @@ import { ColumnId, Datum, ColumnType, IColumn } from 'dash-table/components/Tabl
 import { QuerySyntaxTree } from 'dash-table/syntax-tree';
 import { IConvertedStyle } from 'dash-table/derived/style';
 
+export interface IActiveElement {
+    is_active?: boolean;
+    is_selected?: boolean;
+}
+
 export interface IConditionalElement {
     filter_query?: string;
 }
@@ -29,12 +34,22 @@ export interface IEditableElement {
 }
 
 export type ConditionalBasicFilter = INamedElement & ITypedElement;
-export type ConditionalDataCell = IConditionalElement & IIndexedRowElement & INamedElement & ITypedElement & IEditableElement;
+export type ConditionalDataCell = IActiveElement & IConditionalElement & IIndexedRowElement & INamedElement & ITypedElement & IEditableElement;
 export type ConditionalCell = INamedElement & ITypedElement;
 export type ConditionalHeader = IIndexedHeaderElement & INamedElement & ITypedElement;
 
 function ifAstFilter(ast: QuerySyntaxTree, datum: Datum) {
     return ast.isValid && ast.evaluate(datum);
+}
+
+export function ifColumnActive(condition: IActiveElement | undefined, active: boolean, selected: boolean) {
+    if (!condition) {
+        return true;
+    }
+
+    // Apply active if active, apply selected if selected AND not active (active and selected styling don't overlap)
+    return (condition.is_active === undefined || active === condition.is_active) &&
+        (condition.is_selected === undefined || (selected === condition.is_selected && !active));
 }
 
 export function ifColumnId(condition: INamedElement | undefined, columnId: ColumnId) {
@@ -97,13 +112,20 @@ export function ifEditable(condition: IEditableElement | undefined, isEditable: 
 
 export type Filter<T> = (s: T[]) => T[];
 
-export const matchesDataCell = (datum: Datum, i: number, column: IColumn): Filter<IConvertedStyle> => R.filter<IConvertedStyle>((style =>
+export const matchesDataCell = (
+    datum: Datum,
+    i: number, column: IColumn,
+    active: boolean,
+    selected: boolean
+): Filter<IConvertedStyle> => R.filter<IConvertedStyle>((style =>
+    style.matchesActive(active, selected) &&
     style.matchesRow(i) &&
     style.matchesColumn(column) &&
     style.matchesFilter(datum)
 ));
 
 export const matchesFilterCell = (column: IColumn): Filter<IConvertedStyle> => R.filter<IConvertedStyle>((style =>
+    !style.checksActive() &&
     style.matchesColumn(column)
 ));
 
@@ -113,6 +135,7 @@ export const matchesHeaderCell = (i: number, column: IColumn): Filter<IConverted
 ));
 
 export const matchesDataOpCell = (datum: Datum, i: number): Filter<IConvertedStyle> => R.filter<IConvertedStyle>((style =>
+    !style.checksActive() &&
     !style.checksColumn() &&
     style.matchesRow(i) &&
     style.matchesFilter(datum)