+++ src/dash-table/derived/cell/wrapperStyles.ts
@@ -2,9 +2,11 @@ import * as R from 'ramda';
 import { CSSProperties } from 'react';
 
 import { memoizeOneFactory } from 'core/memoizer';
-import { Data, Columns, IViewportOffset, SelectedCells } from 'dash-table/components/Table/props';
+import { Data, Columns, IViewportOffset, SelectedCells, ICellCoordinates } from 'dash-table/components/Table/props';
 import { IConvertedStyle, getDataCellStyle, getDataOpCellStyle } from '../style';
 import { traverseMap2, shallowClone } from 'core/math/matrixZipMap';
+
+import isActiveCell from 'dash-table/derived/cell/isActive';
 
 const SELECTED_CELL_STYLE = { backgroundColor: 'var(--selected-background)' };
 
@@ -16,28 +18,39 @@ const partialGetter = (
 ) => traverseMap2(
     data,
     columns,
-    (datum, column, i) => getDataCellStyle(datum, i + offset.rows, column)(styles)
+    (datum, column, i) => getDataCellStyle(datum, i + offset.rows, column, false, false)(styles)
 );
 
 const getter = (
-    styles: CSSProperties[][],
+    baseline: CSSProperties[][],
+    columns: Columns,
+    styles: IConvertedStyle[],
+    data: Data,
     offset: IViewportOffset,
+    activeCell: ICellCoordinates,
     selectedCells: SelectedCells
 ) => {
-    styles = shallowClone(styles);
+    baseline = shallowClone(baseline);
 
     R.forEach(({ row: i, column: j }) => {
-        i -= offset.rows;
-        j -= offset.columns;
+        const iNoOffset = i - offset.rows;
+        const jNoOffset = j - offset.columns;
 
-        if (i < 0 || j < 0 || styles.length <= i || styles[i].length <= j) {
+        if (iNoOffset < 0 || jNoOffset < 0 || baseline.length <= iNoOffset || baseline[iNoOffset].length <= jNoOffset) {
             return;
         }
 
-        styles[i][j] = R.merge(styles[i][j], SELECTED_CELL_STYLE);
+        const active = isActiveCell(activeCell, i, j);
+
+        const style = {
+            ...SELECTED_CELL_STYLE,
+            ...getDataCellStyle(data[i], i + offset.rows, columns[j], active, true)(styles)
+        };
+
+        baseline[iNoOffset][jNoOffset] = style;
     }, selectedCells);
 
-    return styles;
+    return baseline;
 };
 
 const opGetter = (